requirements.txt
fastapi
uvicorn
kafka-python
python-dotenv
pydantic
qdrant-client
sentence-transformers
numpy